+++ tests/testthat/test-silent.R
@@ -113,6 +113,7 @@ test_that("parse_setup is silent", {
   expect_silent(parse_setup(mtf_1991_sps_name))
   expect_silent(parse_setup(mtf_1992_sps_name))
   expect_silent(parse_setup(mtf_1979_sps_name))
+  expect_silent(parse_setup(ca_sedd_2005_ahal_sps_name))
 
 
   expect_silent(parse_setup(missing_value_no_s_name))