+++ NEWS.md
@@ -3,6 +3,7 @@
 * Fix bug caused by `vroom` not handling embedded nulls.
 * Fix bug where code would run even if start column number > end column number, but would crash code later.
 * Fix bug where code wouldn't run if .sps setup file read MISSING VALUE rather than MISSING VALUES.
+* Fix bug when value label is "X thru highest value". Now will change all values >= X to be NA. 
 
 
 # asciiSetupReader 2.3.0